Some cleanup, refactoring & docs
